fix: Move ChannelIcon to common export
