+++ packages/client/components/common/index.tsx
@@ -1,5 +1,5 @@
+export * from "./Channels";
 export * from "./Device";
-export { getChannelIcon } from "./lib/channels";
 export { debounce } from "./lib/debounce";
 export { default as CONFIGURATION } from "./lib/env";
 export { insecureUniqueId } from "./lib/unique";
