+++ packages/client/src/interface/navigation/channels/HomeSidebar.tsx
@@ -8,24 +8,19 @@ import { styled } from "styled-system/jsx";
 
 import { ChannelContextMenu, UserContextMenu } from "@revolt/app";
 import { useClient } from "@revolt/client";
-import { useDevice } from "@revolt/common";
+import { ChannelIcon, useDevice } from "@revolt/common";
 import Instance from "@revolt/instance/Instance";
 import { TextWithEmoji } from "@revolt/markdown";
 import { useModals } from "@revolt/modal";
 import { useLocation, useNavigate } from "@revolt/routing";
 import {
-  Avatar,
   Deferred,
   MenuButton,
   OverflowingText,
   Tooltip,
-  UserStatus,
-  iconSize,
   typography,
 } from "@revolt/ui";
 import { Symbol } from "@revolt/ui/components/utils/Symbol";
-
-import MdClose from "@material-design-icons/svg/outlined/close.svg?component-solid";
 
 import { SidebarBase } from "./common";
 
@@ -310,31 +305,7 @@ function Entry(
               ? "active"
               : "normal"
       }
-      icon={
-        <Switch>
-          <Match when={local.channel.type === "Group"}>
-            <Avatar
-              size={32}
-              shape="rounded-square"
-              fallback={local.channel.name}
-              src={local.channel.iconURL}
-              primaryContrast
-            />
-          </Match>
-          <Match when={local.channel.type === "DirectMessage"}>
-            <Avatar
-              size={32}
-              src={local.channel.iconURL}
-              holepunch="bottom-right"
-              overlay={
-                <UserStatus.Graphic
-                  status={local.channel?.recipient?.presence}
-                />
-              }
-            />
-          </Match>
-        </Switch>
-      }
+      icon={<ChannelIcon channel={props.channel} />}
       actions={
         <Show when={!local.isMobile}>
           <a
@@ -346,7 +317,7 @@ function Entry(
               });
             }}
           >
-            <MdClose {...iconSize("18px")} />
+            <Symbol size={18}>close</Symbol>
           </a>
         </Show>
       }
