+++ packages/client/src/interface/navigation/channels/ServerSidebar.tsx
@@ -13,7 +13,7 @@ import { useLingui } from "@lingui/solid/macro";
 import type { API, Channel, Server, ServerFlags } from "stoat.js";
 import { styled } from "styled-system/jsx";
 
-import { getChannelIcon, useDevice } from "@revolt/common";
+import { ChannelIcon, useDevice } from "@revolt/common";
 import { KeybindAction, createKeybind } from "@revolt/keybinds";
 import { TextWithEmoji } from "@revolt/markdown";
 import { useModals } from "@revolt/modal";
@@ -547,82 +547,3 @@ function Entry(
     </Column>
   );
 }
-
-export function ChannelIcon(props: { channel: Channel }) {
-  const voice = useVoice();
-  const inCall = () => props.channel.id === voice.channel()?.id;
-
-  return (
-    <>
-      <Icon
-        channel={props.channel}
-        symbol={getChannelIcon(props.channel)}
-        special={props.channel.mature ? "warning" : undefined}
-        color={inCall() ? "var(--md-sys-color-primary)" : undefined}
-      />
-      <Show when={props.channel.icon}>
-        <CustomIcon src={props.channel.iconURL} css={{ marginEnd: ".2em" }} />
-      </Show>
-    </>
-  );
-}
-
-function Icon(props: {
-  channel: Channel;
-  symbol: string;
-  special?: string;
-  color?: string;
-}) {
-  const maskId = "_m" + Date.now();
-  return (
-    <Show
-      when={props.special}
-      fallback={<Symbol color={props.color}>{props.symbol}</Symbol>}
-    >
-      <svg viewBox="0 0 24 24" width="24" height="24">
-        <mask id={maskId}>
-          <rect x="0" y="0" width="24" height="24" fill="#fff" />
-          <text {...warnProps} stroke="#000" stroke-width="3" color="#000">
-            {props.special}
-          </text>
-        </mask>
-        <text
-          class={"material-symbols-outlined"}
-          style={{
-            "font-variation-settings": '"FILL" 0, "wght" 400, "GRAD" 0',
-            color: props.color,
-          }}
-          x="0"
-          y="24"
-          mask={`url(#${maskId})`}
-        >
-          {props.symbol}
-        </text>
-        <text {...warnProps} color={props.color}>
-          {props.special}
-        </text>
-      </svg>
-    </Show>
-  );
-}
-
-const warnProps = {
-  class: "material-symbols-rounded",
-  style: {
-    "font-variation-settings": '"FILL" 0, "wght" 400, "GRAD" 0, "opsz" 24',
-    "font-size": "10px",
-  },
-  x: 13,
-  y: 10,
-};
-
-/**
- * Channel icon styling
- */
-const CustomIcon = styled("img", {
-  base: {
-    width: "16px",
-    height: "16px",
-    objectFit: "contain",
-  },
-});
